feat: Architecture hardening — security, scalability, and code quality
Security:
- Add rate limiting on /auth/login (slowapi, 5/min)
- Add structured JSON audit logging for auth, remediation, and WebSocket events
- Replace in-memory OIDC state dict with pluggable store (memory/SQLite/Redis)
- Add per-user WebSocket connection limits with shared auth extraction

Scalability:
- Split portainer_client.py (1263 lines) into portainer_http, portainer_normalizers, portainer_utils
- Split config.py (655 lines) into config/ package with 11 sub-modules
- Add filelock-based leader election for multi-worker scheduler guard
- Add file cache eviction (expired + LRU strategies)

Code Quality:
- Add async event bus for monitoring service decoupling
- Extend FastAPI dependency injection with new providers
- Extract magic numbers to configurable settings
- Add pytest coverage reporting (50% threshold)
- Cache session status in Streamlit frontend to reduce redundant API calls

Posture:
- Bump dependency versions (cryptography, jinja2, authlib, pydantic, etc.)
- Add pip-audit to dev dependencies
- Update CLAUDE.md with security documentation
