+++ pyproject.toml
@@ -6,22 +6,23 @@ license = { text = "MIT" }
 requires-python = ">=3.12"
 dependencies = [
     "fastapi>=0.115.0",
-    "uvicorn[standard]>=0.32.0",
-    "httpx>=0.28.0",
-    "pydantic>=2.10.0",
-    "pydantic-settings>=2.7.0",
-    "jinja2>=3.1.0",
+    "uvicorn[standard]>=0.34.0",
+    "httpx>=0.28.1",
+    "pydantic>=2.11.0",
+    "pydantic-settings>=2.9.0",
+    "jinja2>=3.1.5",
     "python-multipart>=0.0.18",
-    "authlib>=1.4.0",
+    "authlib>=1.5.0",
     "PyJWT>=2.10.0",
     "apscheduler>=3.11.0",
     "pandas>=2.3.0",
     "tenacity>=9.0.0",
     "filelock>=3.20.0",
     "itsdangerous>=2.2.0",
-    "redis>=5.0.0",
+    "redis>=5.2.0",
+    "slowapi>=0.1.9",
     # Security: Pin cryptography to fix CVE-2023-50782, CVE-2024-0727, PYSEC-2024-225, GHSA-h4gh-qq45-vh27
-    "cryptography>=43.0.1",
+    "cryptography>=44.0.0",
 ]
 
 [project.optional-dependencies]
@@ -41,10 +42,11 @@ dev = [
     "pytest>=8.0.0",
     "pytest-asyncio>=0.24.0",
     "pytest-cov>=6.0.0",
-    "httpx[http2]>=0.28.0",
+    "httpx[http2]>=0.28.1",
     "ruff>=0.8.0",
     "mypy>=1.14.0",
     "pandas-stubs>=2.2.0",
+    "pip-audit>=2.7.0",
 ]
 e2e = [
     "playwright>=1.40.0",
@@ -65,6 +67,7 @@ packages = ["src/portainer_dashboard"]
 testpaths = ["tests"]
 asyncio_mode = "auto"
 asyncio_default_fixture_loop_scope = "function"
+addopts = "--cov=src/portainer_dashboard --cov-report=term-missing --cov-fail-under=50"
 
 [tool.ruff]
 line-length = 100
