+++ src/portainer_dashboard/config/__init__.py
@@ -1,0 +1,54 @@
+"""Application configuration via Pydantic Settings."""
+
+from .auth import (
+    AuthSettings,
+    OIDCSettings,
+    RateLimitSettings,
+    StaticAuthSettings,
+    WebSocketSettings,
+)
+from .cache import CacheSettings
+from .helpers import (
+    ConfigurationError,
+    PROJECT_ROOT,
+    _empty_str_to_default_bool,
+    _empty_str_to_default_int,
+    _empty_str_to_none,
+    _project_root,
+)
+from .llm import LLMSettings
+from .monitoring import (
+    KibanaSettings,
+    MetricsSettings,
+    MonitoringSettings,
+    RemediationSettings,
+)
+from .portainer import PortainerEnvironmentSettings, PortainerSettings
+from .server import ServerSettings
+from .session import SessionSettings
+from .settings import Settings, get_settings, reload_settings
+from .tracing import TracingSettings
+
+__all__ = [
+    "AuthSettings",
+    "CacheSettings",
+    "ConfigurationError",
+    "KibanaSettings",
+    "LLMSettings",
+    "MetricsSettings",
+    "MonitoringSettings",
+    "OIDCSettings",
+    "PortainerEnvironmentSettings",
+    "PortainerSettings",
+    "PROJECT_ROOT",
+    "RateLimitSettings",
+    "RemediationSettings",
+    "SessionSettings",
+    "ServerSettings",
+    "Settings",
+    "StaticAuthSettings",
+    "TracingSettings",
+    "WebSocketSettings",
+    "get_settings",
+    "reload_settings",
+]
