+++ src/portainer_dashboard/dependencies.py
@@ -74,10 +74,38 @@ async def get_template_context(
 TemplateContextDep = Annotated[dict, Depends(get_template_context)]
 
 
+def get_client_pool_dep():
+    """Dependency that provides the Portainer client pool."""
+    from portainer_dashboard.services.portainer_client import get_client_pool
+    return get_client_pool()
+
+
+def get_event_bus_dep():
+    """Dependency that provides the event bus."""
+    from portainer_dashboard.core.event_bus import get_event_bus
+    return get_event_bus()
+
+
+def get_audit_logger_dep():
+    """Dependency that provides the audit logger."""
+    from portainer_dashboard.core.audit import get_audit_logger
+    return get_audit_logger()
+
+
 def reset_dependencies() -> None:
     """Clear all cached dependencies. Useful for testing."""
     _get_session_storage.cache_clear()
     _get_jinja_env.cache_clear()
+
+    from portainer_dashboard.core.audit import reset_audit_logger
+    from portainer_dashboard.core.event_bus import reset_event_bus
+    from portainer_dashboard.core.ws_limiter import reset_ws_tracker
+    from portainer_dashboard.core.oidc_state_store import reset_oidc_state_store
+
+    reset_audit_logger()
+    reset_event_bus()
+    reset_ws_tracker()
+    reset_oidc_state_store()
 
 
 __all__ = [
@@ -85,6 +113,9 @@ __all__ = [
     "SessionStorageDep",
     "SettingsDep",
     "TemplateContextDep",
+    "get_audit_logger_dep",
+    "get_client_pool_dep",
+    "get_event_bus_dep",
     "get_jinja_env",
     "get_session_storage",
     "get_settings_dep",
