+++ src/portainer_dashboard/services/portainer_client.py
@@ -1,1262 +1,17 @@
-"""Async Portainer API client using httpx with connection pooling."""
-
-from __future__ import annotations
-
-import asyncio
-import logging
-import re
-from dataclasses import dataclass, field
-from typing import Any
-from urllib.parse import urlparse
-
-import httpx
-import pandas as pd
-from tenacity import retry, stop_after_attempt, wait_exponential
-
-from portainer_dashboard.config import PortainerEnvironmentSettings, get_settings
-
-LOGGER = logging.getLogger(__name__)
-
-# Connection pool limits for better performance
-_DEFAULT_MAX_CONNECTIONS = 20
-_DEFAULT_MAX_KEEPALIVE_CONNECTIONS = 10
-_DEFAULT_KEEPALIVE_EXPIRY = 30.0  # seconds
-
-
-class PortainerClientPool:
-    """Connection pool manager for Portainer API clients.
-
-    Maintains a pool of httpx.AsyncClient instances for connection reuse,
-    significantly reducing TCP/TLS handshake overhead for repeated requests.
-    """
-
-    def __init__(self) -> None:
-        self._clients: dict[str, httpx.AsyncClient] = {}
-        self._lock = asyncio.Lock()
-
-    async def get_client(
-        self,
-        base_url: str,
-        api_key: str,
-        *,
-        timeout: float = 60.0,
-        verify_ssl: bool = True,
-    ) -> httpx.AsyncClient:
-        """Get or create a pooled client for the given base URL."""
-        # Use base_url as key (API key might change, but URL identifies the server)
-        key = base_url
-
-        async with self._lock:
-            if key not in self._clients:
-                limits = httpx.Limits(
-                    max_connections=_DEFAULT_MAX_CONNECTIONS,
-                    max_keepalive_connections=_DEFAULT_MAX_KEEPALIVE_CONNECTIONS,
-                    keepalive_expiry=_DEFAULT_KEEPALIVE_EXPIRY,
-                )
-                self._clients[key] = httpx.AsyncClient(
-                    base_url=base_url,
-                    headers={"X-API-Key": api_key},
-                    timeout=timeout,
-                    verify=verify_ssl,
-                    limits=limits,
-                )
-                LOGGER.debug("Created pooled client for %s", base_url)
-
-            return self._clients[key]
-
-    async def close_all(self) -> None:
-        """Close all pooled clients."""
-        async with self._lock:
-            for url, client in self._clients.items():
-                try:
-                    await client.aclose()
-                    LOGGER.debug("Closed pooled client for %s", url)
-                except Exception as exc:
-                    LOGGER.warning("Error closing client for %s: %s", url, exc)
-            self._clients.clear()
-
-    async def close_client(self, base_url: str) -> None:
-        """Close a specific client."""
-        async with self._lock:
-            if base_url in self._clients:
-                try:
-                    await self._clients[base_url].aclose()
-                except Exception as exc:
-                    LOGGER.warning("Error closing client for %s: %s", base_url, exc)
-                del self._clients[base_url]
-
-
-# Global client pool singleton
-_client_pool: PortainerClientPool | None = None
-
-
-def get_client_pool() -> PortainerClientPool:
-    """Get or create the global client pool."""
-    global _client_pool
-    if _client_pool is None:
-        _client_pool = PortainerClientPool()
-    return _client_pool
-
-
-async def shutdown_client_pool() -> None:
-    """Shutdown the global client pool. Call during application shutdown."""
-    global _client_pool
-    if _client_pool is not None:
-        await _client_pool.close_all()
-        _client_pool = None
-
-
-class PortainerAPIError(RuntimeError):
-    """Raised when a Portainer API request fails."""
-
-
-def _coerce_int(value: object) -> int | None:
-    """Return value as an integer when possible."""
-    if value is None:
-        return None
-    if isinstance(value, bool):
-        return int(value)
-    if isinstance(value, (int, float)):
-        return int(value)
-    if isinstance(value, str) and value.strip():
-        try:
-            return int(float(value))
-        except ValueError:
-            return None
-    return None
-
-
-def _first_present(mapping: dict[str, object], *keys: str) -> object:
-    """Return the first value present for keys in mapping."""
-    for key in keys:
-        if key not in mapping:
-            continue
-        value = mapping.get(key)
-        if value is not None:
-            return value
-    return None
-
-
-def _stack_targets_endpoint(stack: dict[str, object], endpoint_id: int) -> bool:
-    """Return True when a stack is assigned to the provided endpoint."""
-    endpoint_keys = ("EndpointId", "EndpointID", "endpointId", "endpointID")
-    for key in endpoint_keys:
-        if key not in stack:
-            continue
-        coerced = _coerce_int(stack.get(key))
-        if coerced is None:
-            continue
-        if coerced == endpoint_id:
-            return True
-
-    deployment_info = stack.get("DeploymentInfo") or stack.get("deploymentInfo")
-    if isinstance(deployment_info, dict):
-        str_endpoint_id = str(endpoint_id)
-        if str_endpoint_id in {str(key) for key in deployment_info.keys()}:
-            return True
-        for info in deployment_info.values():
-            if not isinstance(info, dict):
-                continue
-            for key in endpoint_keys:
-                if key not in info:
-                    continue
-                coerced = _coerce_int(info.get(key))
-                if coerced is None:
-                    continue
-                if coerced == endpoint_id:
-                    return True
-
-    return False
-
-
-def _stack_has_endpoint_metadata(stack: dict[str, object]) -> bool:
-    """Return True when the stack embeds any endpoint assignment metadata."""
-    endpoint_keys = ("EndpointId", "EndpointID", "endpointId", "endpointID")
-    for key in endpoint_keys:
-        if key not in stack:
-            continue
-        coerced = _coerce_int(stack.get(key))
-        if coerced is None:
-            continue
-        return True
-
-    deployment_info = stack.get("DeploymentInfo") or stack.get("deploymentInfo")
-    if isinstance(deployment_info, dict):
-        if any(_coerce_int(key) is not None for key in deployment_info.keys()):
-            return True
-        for info in deployment_info.values():
-            if not isinstance(info, dict):
-                continue
-            for key in endpoint_keys:
-                if key not in info:
-                    continue
-                coerced = _coerce_int(info.get(key))
-                if coerced is None:
-                    continue
-                return True
-
-    return False
-
-
-@dataclass
-class AsyncPortainerClient:
-    """Async Portainer API client using httpx with connection pooling.
-
-    When use_pool=True (default), uses shared connection pool for better performance.
-    When use_pool=False, creates a new client per context (legacy behavior).
-    """
-
-    base_url: str
-    api_key: str
-    timeout: float = 60.0  # Increased from 30s to handle slow API responses
-    verify_ssl: bool = True
-    use_pool: bool = True  # Use connection pooling by default
-    _client: httpx.AsyncClient = field(init=False, repr=False)
-    _owns_client: bool = field(init=False, repr=False, default=True)
-
-    def __post_init__(self) -> None:
-        self.base_url = self.base_url.rstrip("/")
-        if not self.base_url.lower().endswith("/api"):
-            self.base_url = f"{self.base_url}/api"
-        if not self.api_key:
-            raise ValueError("Portainer API key is required")
-
-    async def __aenter__(self) -> "AsyncPortainerClient":
-        if self.use_pool:
-            # Use pooled client - don't close on exit
-            pool = get_client_pool()
-            self._client = await pool.get_client(
-                self.base_url,
-                self.api_key,
-                timeout=self.timeout,
-                verify_ssl=self.verify_ssl,
-            )
-            self._owns_client = False
-        else:
-            # Create a new client (legacy behavior)
-            self._client = httpx.AsyncClient(
-                base_url=self.base_url,
-                headers={"X-API-Key": self.api_key},
-                timeout=self.timeout,
-                verify=self.verify_ssl,
-            )
-            self._owns_client = True
-        return self
-
-    async def __aexit__(self, exc_type, exc_val, exc_tb) -> None:
-        # Only close if we own the client (not pooled)
-        if self._owns_client:
-            await self._client.aclose()
-
-    @retry(
-        stop=stop_after_attempt(3),
-        wait=wait_exponential(multiplier=0.5, min=0.5, max=5),
-        reraise=True,
-    )
-    async def _request(
-        self, path: str, *, params: dict[str, object] | None = None
-    ) -> object:
-        try:
-            response = await self._client.get(path, params=params)
-            response.raise_for_status()
-        except httpx.HTTPError as exc:
-            raise PortainerAPIError(str(exc)) from exc
-        try:
-            return response.json()
-        except ValueError as exc:
-            raise PortainerAPIError("Invalid JSON response from Portainer") from exc
-
-    @retry(
-        stop=stop_after_attempt(3),
-        wait=wait_exponential(multiplier=0.5, min=0.5, max=5),
-        reraise=True,
-    )
-    async def _post(
-        self,
-        path: str,
-        *,
-        json: dict[str, object] | None = None,
-    ) -> httpx.Response:
-        try:
-            response = await self._client.post(path, json=json)
-            response.raise_for_status()
-        except httpx.HTTPError as exc:
-            raise PortainerAPIError(str(exc)) from exc
-        return response
-
-    @staticmethod
-    def _extract_filename(header_value: str | None) -> str | None:
-        if not header_value:
-            return None
-        parts = [part.strip() for part in header_value.split(";")]
-        filename: str | None = None
-        for part in parts:
-            if not part:
-                continue
-            key, _, value = part.partition("=")
-            if not value:
-                continue
-            lowered_key = key.strip().lower()
-            cleaned_value = value.strip().strip('"')
-            if lowered_key == "filename*":
-                if cleaned_value.lower().startswith("utf-8''"):
-                    cleaned_value = cleaned_value[7:]
-                import urllib.parse
-                filename = urllib.parse.unquote(cleaned_value)
-                break
-            if lowered_key == "filename":
-                filename = cleaned_value
-        if filename:
-            filename = re.sub(r"[\\/]+", "-", filename).strip()
-        return filename or None
-
-    async def create_backup(
-        self, *, password: str | None = None
-    ) -> tuple[bytes, str | None]:
-        payload: dict[str, object] = {}
-        if password:
-            payload["password"] = password
-        response = await self._post("/backup", json=payload or {})
-        filename = self._extract_filename(response.headers.get("Content-Disposition"))
-        return response.content, filename
-
-    async def list_edge_endpoints(self) -> list[dict[str, object]]:
-        """List only edge endpoints."""
-        params = {"edge": "true", "status": "true"}
-        data = await self._request("/endpoints", params=params)
-        if not isinstance(data, list):
-            raise PortainerAPIError("Unexpected endpoints payload from Portainer")
-        return data
-
-    async def list_all_endpoints(self) -> list[dict[str, object]]:
-        """List all endpoints (including local Docker environments)."""
-        data = await self._request("/endpoints")
-        if not isinstance(data, list):
-            raise PortainerAPIError("Unexpected endpoints payload from Portainer")
-        return data
-
-    async def list_stacks_for_endpoint(
-        self, endpoint_id: int
-    ) -> list[dict[str, object]]:
-        """Fetch stacks for a given endpoint."""
-        paths = (
-            ("/stacks", {"endpointId": endpoint_id}),
-            ("/edge/stacks", {"endpointId": endpoint_id}),
-        )
-        results: list[dict[str, object]] = []
-        seen_stack_ids: set[int] = set()
-
-        for path, params in paths:
-            try:
-                data = await self._request(path, params=params)
-            except PortainerAPIError as exc:
-                LOGGER.debug(
-                    "Failed fetching %s for endpoint %s: %s", path, endpoint_id, exc
-                )
-                continue
-            if not isinstance(data, list):
-                continue
-            for item in data:
-                if not isinstance(item, dict):
-                    continue
-                raw_id = item.get("Id") or item.get("ID") or item.get("id")
-                stack_id = _coerce_int(raw_id)
-                if stack_id is not None and stack_id in seen_stack_ids:
-                    continue
-                if stack_id is not None:
-                    seen_stack_ids.add(stack_id)
-                results.append(item)
-        return results
-
-    async def list_containers_for_endpoint(
-        self,
-        endpoint_id: int,
-        *,
-        include_stopped: bool = False,
-    ) -> list[dict[str, object]]:
-        """Return containers for an endpoint via the Docker API."""
-        params = {"all": "1" if include_stopped else "0"}
-        data = await self._request(
-            f"/endpoints/{endpoint_id}/docker/containers/json",
-            params=params,
-        )
-        if not isinstance(data, list):
-            raise PortainerAPIError("Unexpected containers payload from Portainer")
-        return data
-
-    async def inspect_container(
-        self, endpoint_id: int, container_id: str
-    ) -> dict[str, object]:
-        """Return the detailed inspect payload for a container."""
-        data = await self._request(
-            f"/endpoints/{endpoint_id}/docker/containers/{container_id}/json"
-        )
-        if not isinstance(data, dict):
-            raise PortainerAPIError(
-                "Unexpected container inspect payload from Portainer"
-            )
-        return data
-
-    async def get_container_logs(
-        self,
-        endpoint_id: int,
-        container_id: str,
-        *,
-        tail: int = 500,
-        timestamps: bool = True,
-        since: int | None = None,
-    ) -> str:
-        """Fetch container logs from Docker API via Portainer.
-
-        Args:
-            endpoint_id: The Portainer endpoint ID.
-            container_id: The container ID or name.
-            tail: Number of lines to return from the end of the logs.
-            timestamps: Whether to include timestamps in each log line.
-            since: Only return logs since this Unix timestamp.
-
-        Returns:
-            The container log output as a string.
-        """
-        params = {
-            "stdout": "true",
-            "stderr": "true",
-            "tail": str(tail),
-            "timestamps": "true" if timestamps else "false",
-        }
-        if since is not None:
-            params["since"] = str(since)
-
-        try:
-            response = await self._client.get(
-                f"/endpoints/{endpoint_id}/docker/containers/{container_id}/logs",
-                params=params,
-            )
-            response.raise_for_status()
-        except httpx.HTTPError as exc:
-            raise PortainerAPIError(str(exc)) from exc
-
-        # Docker logs API returns raw text, may include special characters
-        # Strip any Docker stream headers (first 8 bytes per frame)
-        text = response.text
-        # Clean up the output by removing Docker stream header bytes
-        lines = []
-        for line in text.split("\n"):
-            # Docker multiplexed streams have 8-byte headers
-            # Check if line starts with binary header and clean it
-            if len(line) > 8 and line[0] in "\x00\x01\x02":
-                lines.append(line[8:])
-            else:
-                lines.append(line)
-        return "\n".join(lines)
-
-    async def get_container_stats(
-        self, endpoint_id: int, container_id: str
-    ) -> dict[str, object]:
-        """Return a non-streaming stats snapshot for a container."""
-        data = await self._request(
-            f"/endpoints/{endpoint_id}/docker/containers/{container_id}/stats",
-            params={"stream": "false"},
-        )
-        if not isinstance(data, dict):
-            raise PortainerAPIError("Unexpected container stats payload from Portainer")
-        return data
-
-    async def get_endpoint_host_info(self, endpoint_id: int) -> dict[str, object]:
-        """Return Docker host metadata for an endpoint."""
-        data = await self._request(f"/endpoints/{endpoint_id}/docker/info")
-        if not isinstance(data, dict):
-            raise PortainerAPIError("Unexpected host info payload from Portainer")
-        return data
-
-    async def get_endpoint_system_df(self, endpoint_id: int) -> dict[str, object]:
-        """Return Docker disk usage statistics for an endpoint."""
-        data = await self._request(f"/endpoints/{endpoint_id}/docker/system/df")
-        if not isinstance(data, dict):
-            raise PortainerAPIError("Unexpected system df payload from Portainer")
-        return data
-
-    async def list_volumes_for_endpoint(
-        self, endpoint_id: int
-    ) -> list[dict[str, object]]:
-        """Return all Docker volumes defined on an endpoint."""
-        data = await self._request(f"/endpoints/{endpoint_id}/docker/volumes")
-        if isinstance(data, dict):
-            volumes = data.get("Volumes")
-        else:
-            volumes = None
-        if volumes is None:
-            return []
-        if not isinstance(volumes, list):
-            raise PortainerAPIError("Unexpected volumes payload from Portainer")
-        return [item for item in volumes if isinstance(item, dict)]
-
-    async def list_images_for_endpoint(
-        self, endpoint_id: int
-    ) -> list[dict[str, object]]:
-        """Return image metadata for an endpoint."""
-        data = await self._request(f"/endpoints/{endpoint_id}/docker/images/json")
-        if not isinstance(data, list):
-            raise PortainerAPIError("Unexpected images payload from Portainer")
-        return [item for item in data if isinstance(item, dict)]
-
-    async def get_stack_image_status(self, stack_id: int) -> object:
-        """Return the image status payload for the specified stack."""
-        data = await self._request(f"/stacks/{stack_id}/images_status")
-        if not isinstance(data, (dict, list)):
-            raise PortainerAPIError(
-                "Unexpected stack image status payload from Portainer"
-            )
-        return data
-
-    async def restart_container(
-        self, endpoint_id: int, container_id: str, *, timeout: int = 10
-    ) -> dict[str, object]:
-        """Restart a container.
-
-        Args:
-            endpoint_id: The Portainer endpoint ID.
-            container_id: The container ID or name.
-            timeout: Seconds to wait before killing the container.
-
-        Returns:
-            A dict with success status.
-        """
-        try:
-            response = await self._client.post(
-                f"/endpoints/{endpoint_id}/docker/containers/{container_id}/restart",
-                params={"t": str(timeout)},
-            )
-            response.raise_for_status()
-        except httpx.HTTPError as exc:
-            raise PortainerAPIError(f"Failed to restart container: {exc}") from exc
-        return {"success": True, "action": "restart", "container_id": container_id}
-
-    async def start_container(
-        self, endpoint_id: int, container_id: str
-    ) -> dict[str, object]:
-        """Start a stopped container.
-
-        Args:
-            endpoint_id: The Portainer endpoint ID.
-            container_id: The container ID or name.
-
-        Returns:
-            A dict with success status.
-        """
-        try:
-            response = await self._client.post(
-                f"/endpoints/{endpoint_id}/docker/containers/{container_id}/start",
-            )
-            # 304 means container is already started
-            if response.status_code == 304:
-                return {"success": True, "action": "start", "container_id": container_id, "note": "already running"}
-            response.raise_for_status()
-        except httpx.HTTPError as exc:
-            raise PortainerAPIError(f"Failed to start container: {exc}") from exc
-        return {"success": True, "action": "start", "container_id": container_id}
-
-    async def stop_container(
-        self, endpoint_id: int, container_id: str, *, timeout: int = 10
-    ) -> dict[str, object]:
-        """Stop a running container.
-
-        Args:
-            endpoint_id: The Portainer endpoint ID.
-            container_id: The container ID or name.
-            timeout: Seconds to wait before killing the container.
-
-        Returns:
-            A dict with success status.
-        """
-        try:
-            response = await self._client.post(
-                f"/endpoints/{endpoint_id}/docker/containers/{container_id}/stop",
-                params={"t": str(timeout)},
-            )
-            # 304 means container is already stopped
-            if response.status_code == 304:
-                return {"success": True, "action": "stop", "container_id": container_id, "note": "already stopped"}
-            response.raise_for_status()
-        except httpx.HTTPError as exc:
-            raise PortainerAPIError(f"Failed to stop container: {exc}") from exc
-        return {"success": True, "action": "stop", "container_id": container_id}
-
-
-def create_portainer_client(
-    env: PortainerEnvironmentSettings,
-) -> AsyncPortainerClient:
-    """Create an async Portainer client for the given environment."""
-    return AsyncPortainerClient(
-        base_url=env.api_url,
-        api_key=env.api_key,
-        verify_ssl=env.verify_ssl,
-        timeout=env.timeout,
-    )
-
-
-# Data normalizers (kept from original implementation)
-
-
-def _determine_edge_agent_status(
-    endpoint: dict[str, object],
-    raw_status: object,
-) -> int:
-    """Determine the actual status for an edge agent.
-
-    Edge agents may report status differently than regular Docker endpoints.
-    This function checks multiple indicators to determine if an edge agent
-    is actually online or offline.
-
-    Returns:
-        1 for Online, 2 for Offline
-    """
-    import time
-
-    # Check endpoint type - Type 4 is Edge Agent, Type 7 is Edge Agent (async)
-    endpoint_type = _coerce_int(_first_present(endpoint, "Type", "type"))
-    is_edge_agent = endpoint_type in (4, 7)
-
-    # If raw status is valid and it's not an edge agent, trust it
-    if not is_edge_agent:
-        if raw_status == 1:
-            return 1
-        if raw_status == 2:
-            return 2
-        # For non-edge, unknown status defaults to offline
-        return 2
-
-    # For edge agents, check the heartbeat/check-in time
-    # EdgeCheckinInterval is in seconds
-    checkin_interval = _coerce_int(
-        _first_present(endpoint, "EdgeCheckinInterval", "edgeCheckinInterval")
-    )
-
-    # Get the last check-in timestamp
-    last_checkin = _first_present(
-        endpoint,
-        "LastCheckInDate",
-        "EdgeLastCheckInDate",
-        "LastCheckIn",
-        "lastCheckInDate",
-    )
-
-    # Convert to Unix timestamp if possible
-    last_checkin_ts: float | None = None
-    if isinstance(last_checkin, (int, float)) and last_checkin > 0:
-        last_checkin_ts = float(last_checkin)
-    elif isinstance(last_checkin, str) and last_checkin:
-        try:
-            parsed = pd.to_datetime(last_checkin, utc=True)
-            if not pd.isna(parsed):
-                last_checkin_ts = parsed.timestamp()
-        except (ValueError, TypeError):
-            pass
-
-    # If we have a last check-in time, use it to determine status
-    if last_checkin_ts is not None:
-        now = time.time()
-        time_since_checkin = now - last_checkin_ts
-
-        # Default check-in interval for edge agents is typically 5 seconds
-        # Consider offline if no check-in for 3x the interval (or 60 seconds if unknown)
-        expected_interval = checkin_interval if checkin_interval and checkin_interval > 0 else 5
-        offline_threshold = max(expected_interval * 3, 60)
-
-        if time_since_checkin <= offline_threshold:
-            return 1  # Online
-        else:
-            return 2  # Offline
-
-    # Check the Heartbeat field if available
-    heartbeat = _first_present(endpoint, "Heartbeat", "heartbeat")
-    if heartbeat is True or heartbeat == 1:
-        return 1
-
-    # Fall back to raw status
-    if raw_status == 1:
-        return 1
-    if raw_status == 2:
-        return 2
-
-    # Edge agents with no status info default to offline
-    return 2
-
-
-def normalise_endpoint_metadata(
-    endpoints: list[dict[str, object]]
-) -> pd.DataFrame:
-    """Return a dataframe with enriched endpoint metadata."""
-    records: list[dict[str, object]] = []
-    for endpoint in endpoints:
-        endpoint_id = int(_first_present(endpoint, "Id", "id") or 0)
-        endpoint_name = endpoint.get("Name") or endpoint.get("name")
-        raw_status = _first_present(endpoint, "Status", "status")
-
-        # Use enhanced status detection for edge agents
-        endpoint_status = _determine_edge_agent_status(endpoint, raw_status)
-
-        agent = endpoint.get("Agent") or endpoint.get("agent") or {}
-        if not isinstance(agent, dict):
-            agent = {}
-        agent_version = agent.get("Version") or agent.get("version")
-        platform = agent.get("Platform") or agent.get("platform")
-        os_name = agent.get("Os") or agent.get("OS") or agent.get("os")
-        group_id = _coerce_int(
-            _first_present(endpoint, "GroupId", "GroupID", "groupId")
-        )
-        tags = endpoint.get("Tags") or endpoint.get("tags")
-        if isinstance(tags, list):
-            tag_summary = ", ".join(str(tag) for tag in tags if tag is not None)
-        else:
-            tag_summary = str(tags) if tags not in (None, "") else None
-        last_check_in = _first_present(
-            endpoint,
-            "LastCheckInDate",
-            "EdgeLastCheckInDate",
-            "LastCheckIn",
-        )
-        last_check_iso: str | None = None
-        if isinstance(last_check_in, (int, float)):
-            try:
-                last_check_iso = pd.to_datetime(
-                    last_check_in, unit="s", utc=True
-                ).isoformat()
-            except (ValueError, OverflowError):
-                last_check_iso = None
-        elif isinstance(last_check_in, str) and last_check_in:
-            try:
-                last_check_iso = pd.to_datetime(last_check_in, utc=True).isoformat()
-            except (ValueError, TypeError):
-                last_check_iso = last_check_in
-        url_value = endpoint.get("URL") or endpoint.get("Url") or endpoint.get("url")
-
-        def _parse_hostname(value: object) -> str | None:
-            if not isinstance(value, str):
-                return None
-            candidate = value.strip()
-            if not candidate:
-                return None
-            parsed = urlparse(
-                candidate if "://" in candidate else f"tcp://{candidate}"
-            )
-            hostname = parsed.hostname
-            if hostname:
-                return hostname
-            if ":" in candidate and "//" not in candidate:
-                return candidate.split(":", 1)[0]
-            return candidate or None
-
-        agent_hostname = _parse_hostname(url_value)
-        if not agent_hostname:
-            public_url = (
-                endpoint.get("PublicURL")
-                or endpoint.get("PublicUrl")
-                or endpoint.get("publicURL")
-                or endpoint.get("publicUrl")
-                or endpoint.get("public_url")
-            )
-            agent_hostname = _parse_hostname(public_url)
-
-        record = {
-            "endpoint_id": endpoint_id,
-            "endpoint_name": endpoint_name,
-            "endpoint_status": endpoint_status,
-            "agent_version": agent_version,
-            "platform": platform,
-            "operating_system": os_name,
-            "group_id": group_id,
-            "tags": tag_summary,
-            "last_check_in": last_check_iso,
-            "url": url_value,
-            "agent_hostname": agent_hostname,
-        }
-        records.append(record)
-    if not records:
-        return pd.DataFrame(
-            columns=[
-                "endpoint_id",
-                "endpoint_name",
-                "endpoint_status",
-                "agent_version",
-                "platform",
-                "operating_system",
-                "group_id",
-                "tags",
-                "last_check_in",
-                "url",
-                "agent_hostname",
-            ]
-        )
-    return pd.DataFrame.from_records(records)
-
-
-def normalise_endpoint_stacks(
-    endpoints: list[dict[str, object]],
-    stacks_by_endpoint: dict[int, list[dict[str, object]]],
-) -> pd.DataFrame:
-    """Return a normalised dataframe mapping endpoints to stacks."""
-    records: list[dict[str, object]] = []
-    for endpoint in endpoints:
-        endpoint_id = int(_first_present(endpoint, "Id", "id") or 0)
-        endpoint_name = endpoint.get("Name") or endpoint.get("name")
-        endpoint_status = _first_present(endpoint, "Status", "status")
-        raw_stacks = stacks_by_endpoint.get(endpoint_id, [])
-        targeted_stacks = [
-            stack
-            for stack in raw_stacks
-            if _stack_targets_endpoint(stack, endpoint_id)
-        ]
-        stacks = targeted_stacks
-        if not stacks:
-            stacks = [
-                stack
-                for stack in raw_stacks
-                if not _stack_has_endpoint_metadata(stack)
-            ]
-
-        if stacks:
-            for stack in stacks:
-                records.append(
-                    {
-                        "endpoint_id": endpoint_id,
-                        "endpoint_name": endpoint_name,
-                        "endpoint_status": endpoint_status,
-                        "stack_id": _first_present(stack, "Id", "id"),
-                        "stack_name": stack.get("Name") or stack.get("name"),
-                        "stack_status": _first_present(stack, "Status", "status"),
-                        "stack_type": _first_present(stack, "Type", "type"),
-                    }
-                )
-        else:
-            records.append(
-                {
-                    "endpoint_id": endpoint_id,
-                    "endpoint_name": endpoint_name,
-                    "endpoint_status": endpoint_status,
-                    "stack_id": None,
-                    "stack_name": None,
-                    "stack_status": None,
-                    "stack_type": None,
-                }
-            )
-    if not records:
-        return pd.DataFrame(
-            columns=[
-                "endpoint_id",
-                "endpoint_name",
-                "endpoint_status",
-                "stack_id",
-                "stack_name",
-                "stack_status",
-                "stack_type",
-            ]
-        )
-    return pd.DataFrame.from_records(records)
-
-
-def normalise_endpoint_containers(
-    endpoints: list[dict[str, object]],
-    containers_by_endpoint: dict[int, list[dict[str, object]]],
-) -> pd.DataFrame:
-    """Return a normalised dataframe mapping endpoints to containers."""
-    records: list[dict[str, object]] = []
-    for endpoint in endpoints:
-        endpoint_id = int(_first_present(endpoint, "Id", "id") or 0)
-        endpoint_name = endpoint.get("Name") or endpoint.get("name")
-        containers = containers_by_endpoint.get(endpoint_id, [])
-        for container in containers:
-            names = container.get("Names") or []
-            if isinstance(names, list) and names:
-                container_name = str(names[0]).lstrip("/")
-            else:
-                container_name = container.get("Name") or container.get("name")
-            image = container.get("Image") or container.get("ImageID")
-            state = container.get("State")
-            status = container.get("Status")
-            restart_count = container.get("RestartCount")
-            created_raw = container.get("Created")
-            created_at: str | None = None
-            if isinstance(created_raw, (int, float)):
-                created_at = pd.to_datetime(
-                    created_raw, unit="s", utc=True
-                ).isoformat()
-            elif isinstance(created_raw, str):
-                try:
-                    created_at_ts = pd.to_datetime(created_raw, utc=True)
-                except (TypeError, ValueError):
-                    created_at_ts = pd.NaT
-                created_at = (
-                    created_at_ts.isoformat()
-                    if isinstance(created_at_ts, pd.Timestamp)
-                    else created_raw
-                )
-            ports = container.get("Ports")
-            port_summary = None
-            if isinstance(ports, list) and ports:
-                summaries = []
-                for port in ports:
-                    private_port = port.get("PrivatePort")
-                    public_port = port.get("PublicPort")
-                    type_ = port.get("Type")
-                    if private_port is None:
-                        continue
-                    if public_port:
-                        summaries.append(
-                            f"{public_port}->{private_port}/{type_}"
-                            if type_
-                            else f"{public_port}->{private_port}"
-                        )
-                    else:
-                        summaries.append(
-                            f"{private_port}/{type_}" if type_ else str(private_port)
-                        )
-                if summaries:
-                    port_summary = ", ".join(summaries)
-            records.append(
-                {
-                    "endpoint_id": endpoint_id,
-                    "endpoint_name": endpoint_name,
-                    "container_id": container.get("Id")
-                    or container.get("ID")
-                    or container.get("id"),
-                    "container_name": container_name,
-                    "image": image,
-                    "state": state,
-                    "status": status,
-                    "restart_count": restart_count,
-                    "created_at": created_at,
-                    "ports": port_summary,
-                }
-            )
-    if not records:
-        return pd.DataFrame(
-            columns=[
-                "endpoint_id",
-                "endpoint_name",
-                "container_id",
-                "container_name",
-                "image",
-                "state",
-                "status",
-                "restart_count",
-                "created_at",
-                "ports",
-            ]
-        )
-    return pd.DataFrame.from_records(records)
-
-
-def normalise_endpoint_images(
-    endpoints: list[dict[str, object]],
-    images_by_endpoint: dict[int, list[dict[str, object]]],
-) -> pd.DataFrame:
-    """Return normalised image data for endpoints."""
-    records: list[dict[str, object]] = []
-    endpoint_lookup = {
-        int(_first_present(endpoint, "Id", "id") or 0): endpoint
-        for endpoint in endpoints
-    }
-    for endpoint_id, images in images_by_endpoint.items():
-        endpoint = endpoint_lookup.get(endpoint_id, {})
-        endpoint_name = endpoint.get("Name") or endpoint.get("name")
-        for image in images:
-            if not isinstance(image, dict):
-                continue
-            repo_tags = image.get("RepoTags")
-            if isinstance(repo_tags, list) and repo_tags:
-                reference = repo_tags[0]
-            else:
-                reference = image.get("RepoDigests")
-                if isinstance(reference, list) and reference:
-                    reference = reference[0]
-            if isinstance(reference, list):
-                reference = reference[0] if reference else None
-            created_raw = image.get("Created")
-            created_at = None
-            if isinstance(created_raw, (int, float)):
-                try:
-                    created_at = pd.to_datetime(
-                        created_raw, unit="s", utc=True
-                    ).isoformat()
-                except (OverflowError, ValueError):
-                    created_at = None
-            elif isinstance(created_raw, str) and created_raw:
-                try:
-                    created_at = pd.to_datetime(created_raw, utc=True).isoformat()
-                except (ValueError, TypeError):
-                    created_at = created_raw
-            size = image.get("Size") or image.get("VirtualSize")
-            records.append(
-                {
-                    "endpoint_id": endpoint_id,
-                    "endpoint_name": endpoint_name,
-                    "image_id": image.get("Id") or image.get("ID"),
-                    "reference": reference,
-                    "size": size,
-                    "created_at": created_at,
-                    "dangling": image.get("Dangling"),
-                }
-            )
-    if not records:
-        return pd.DataFrame(
-            columns=[
-                "endpoint_id",
-                "endpoint_name",
-                "image_id",
-                "reference",
-                "size",
-                "created_at",
-                "dangling",
-            ]
-        )
-    return pd.DataFrame.from_records(records)
-
-
-def normalise_endpoint_metadata_dict(
-    endpoints: list[dict[str, object]]
-) -> list[dict[str, object]]:
-    """Return a list of dicts with enriched endpoint metadata.
-
-    This is a dict-based alternative to normalise_endpoint_metadata() that
-    avoids pandas DataFrame overhead when the data will be serialized to
-    dicts anyway (e.g., for caching).
-    """
-    records: list[dict[str, object]] = []
-    for endpoint in endpoints:
-        endpoint_id = int(_first_present(endpoint, "Id", "id") or 0)
-        endpoint_name = endpoint.get("Name") or endpoint.get("name")
-        raw_status = _first_present(endpoint, "Status", "status")
-
-        # Use enhanced status detection for edge agents
-        endpoint_status = _determine_edge_agent_status(endpoint, raw_status)
-
-        agent = endpoint.get("Agent") or endpoint.get("agent") or {}
-        if not isinstance(agent, dict):
-            agent = {}
-        agent_version = agent.get("Version") or agent.get("version")
-        platform = agent.get("Platform") or agent.get("platform")
-        os_name = agent.get("Os") or agent.get("OS") or agent.get("os")
-        group_id = _coerce_int(
-            _first_present(endpoint, "GroupId", "GroupID", "groupId")
-        )
-        tags = endpoint.get("Tags") or endpoint.get("tags")
-        if isinstance(tags, list):
-            tag_summary = ", ".join(str(tag) for tag in tags if tag is not None)
-        else:
-            tag_summary = str(tags) if tags not in (None, "") else None
-        last_check_in = _first_present(
-            endpoint,
-            "LastCheckInDate",
-            "EdgeLastCheckInDate",
-            "LastCheckIn",
-        )
-        last_check_iso: str | None = None
-        if isinstance(last_check_in, (int, float)):
-            try:
-                from datetime import datetime, timezone
-                last_check_iso = datetime.fromtimestamp(
-                    last_check_in, tz=timezone.utc
-                ).isoformat()
-            except (ValueError, OverflowError, OSError):
-                last_check_iso = None
-        elif isinstance(last_check_in, str) and last_check_in:
-            # Keep string as-is if already formatted
-            last_check_iso = last_check_in
-        url_value = endpoint.get("URL") or endpoint.get("Url") or endpoint.get("url")
-
-        def _parse_hostname(value: object) -> str | None:
-            if not isinstance(value, str):
-                return None
-            candidate = value.strip()
-            if not candidate:
-                return None
-            parsed = urlparse(
-                candidate if "://" in candidate else f"tcp://{candidate}"
-            )
-            hostname = parsed.hostname
-            if hostname:
-                return hostname
-            if ":" in candidate and "//" not in candidate:
-                return candidate.split(":", 1)[0]
-            return candidate or None
-
-        agent_hostname = _parse_hostname(url_value)
-        if not agent_hostname:
-            public_url = (
-                endpoint.get("PublicURL")
-                or endpoint.get("PublicUrl")
-                or endpoint.get("publicURL")
-                or endpoint.get("publicUrl")
-                or endpoint.get("public_url")
-            )
-            agent_hostname = _parse_hostname(public_url)
-
-        record = {
-            "endpoint_id": endpoint_id,
-            "endpoint_name": endpoint_name,
-            "endpoint_status": endpoint_status,
-            "agent_version": agent_version,
-            "platform": platform,
-            "operating_system": os_name,
-            "group_id": group_id,
-            "tags": tag_summary,
-            "last_check_in": last_check_iso,
-            "url": url_value,
-            "agent_hostname": agent_hostname,
-        }
-        records.append(record)
-    return records
-
-
-def normalise_endpoint_stacks_dict(
-    endpoints: list[dict[str, object]],
-    stacks_by_endpoint: dict[int, list[dict[str, object]]],
-) -> list[dict[str, object]]:
-    """Return a list of dicts mapping endpoints to stacks.
-
-    This is a dict-based alternative to normalise_endpoint_stacks() that
-    avoids pandas DataFrame overhead when the data will be serialized to
-    dicts anyway (e.g., for caching).
-    """
-    records: list[dict[str, object]] = []
-    for endpoint in endpoints:
-        endpoint_id = int(_first_present(endpoint, "Id", "id") or 0)
-        endpoint_name = endpoint.get("Name") or endpoint.get("name")
-        endpoint_status = _first_present(endpoint, "Status", "status")
-        raw_stacks = stacks_by_endpoint.get(endpoint_id, [])
-        targeted_stacks = [
-            stack
-            for stack in raw_stacks
-            if _stack_targets_endpoint(stack, endpoint_id)
-        ]
-        stacks = targeted_stacks
-        if not stacks:
-            stacks = [
-                stack
-                for stack in raw_stacks
-                if not _stack_has_endpoint_metadata(stack)
-            ]
-
-        if stacks:
-            for stack in stacks:
-                records.append(
-                    {
-                        "endpoint_id": endpoint_id,
-                        "endpoint_name": endpoint_name,
-                        "endpoint_status": endpoint_status,
-                        "stack_id": _first_present(stack, "Id", "id"),
-                        "stack_name": stack.get("Name") or stack.get("name"),
-                        "stack_status": _first_present(stack, "Status", "status"),
-                        "stack_type": _first_present(stack, "Type", "type"),
-                    }
-                )
-        else:
-            records.append(
-                {
-                    "endpoint_id": endpoint_id,
-                    "endpoint_name": endpoint_name,
-                    "endpoint_status": endpoint_status,
-                    "stack_id": None,
-                    "stack_name": None,
-                    "stack_status": None,
-                    "stack_type": None,
-                }
-            )
-    return records
-
-
-def normalise_endpoint_containers_dict(
-    endpoints: list[dict[str, object]],
-    containers_by_endpoint: dict[int, list[dict[str, object]]],
-) -> list[dict[str, object]]:
-    """Return a list of dicts mapping endpoints to containers.
-
-    This is a dict-based alternative to normalise_endpoint_containers() that
-    avoids pandas DataFrame overhead when the data will be serialized to
-    dicts anyway (e.g., for caching).
-    """
-    records: list[dict[str, object]] = []
-    for endpoint in endpoints:
-        endpoint_id = int(_first_present(endpoint, "Id", "id") or 0)
-        endpoint_name = endpoint.get("Name") or endpoint.get("name")
-        containers = containers_by_endpoint.get(endpoint_id, [])
-        for container in containers:
-            names = container.get("Names") or []
-            if isinstance(names, list) and names:
-                container_name = str(names[0]).lstrip("/")
-            else:
-                container_name = container.get("Name") or container.get("name")
-            image = container.get("Image") or container.get("ImageID")
-            state = container.get("State")
-            status = container.get("Status")
-            restart_count = container.get("RestartCount")
-            created_raw = container.get("Created")
-            created_at: str | None = None
-            if isinstance(created_raw, (int, float)):
-                try:
-                    from datetime import datetime, timezone
-                    created_at = datetime.fromtimestamp(
-                        created_raw, tz=timezone.utc
-                    ).isoformat()
-                except (ValueError, OverflowError, OSError):
-                    created_at = None
-            elif isinstance(created_raw, str):
-                created_at = created_raw
-            ports = container.get("Ports")
-            port_summary = None
-            if isinstance(ports, list) and ports:
-                summaries = []
-                for port in ports:
-                    private_port = port.get("PrivatePort")
-                    public_port = port.get("PublicPort")
-                    type_ = port.get("Type")
-                    if private_port is None:
-                        continue
-                    if public_port:
-                        summaries.append(
-                            f"{public_port}->{private_port}/{type_}"
-                            if type_
-                            else f"{public_port}->{private_port}"
-                        )
-                    else:
-                        summaries.append(
-                            f"{private_port}/{type_}" if type_ else str(private_port)
-                        )
-                if summaries:
-                    port_summary = ", ".join(summaries)
-            records.append(
-                {
-                    "endpoint_id": endpoint_id,
-                    "endpoint_name": endpoint_name,
-                    "container_id": container.get("Id")
-                    or container.get("ID")
-                    or container.get("id"),
-                    "container_name": container_name,
-                    "image": image,
-                    "state": state,
-                    "status": status,
-                    "restart_count": restart_count,
-                    "created_at": created_at,
-                    "ports": port_summary,
-                }
-            )
-    return records
-
-
-__all__ = [
-    "AsyncPortainerClient",
-    "PortainerAPIError",
-    "PortainerClientPool",
-    "_determine_edge_agent_status",
-    "create_portainer_client",
-    "get_client_pool",
-    "normalise_endpoint_containers",
-    "normalise_endpoint_containers_dict",
-    "normalise_endpoint_images",
-    "normalise_endpoint_metadata",
-    "normalise_endpoint_metadata_dict",
-    "normalise_endpoint_stacks",
-    "normalise_endpoint_stacks_dict",
-    "shutdown_client_pool",
-]
+"""Async Portainer API client -- backward-compatible re-exports.
+
+This module has been split into:
+- portainer_http: Client, pool, error classes
+- portainer_normalizers: DataFrame/dict normalisation functions
+- portainer_utils: Internal utility functions
+"""
+from portainer_dashboard.services.portainer_http import *  # noqa: F401,F403
+from portainer_dashboard.services.portainer_normalizers import *  # noqa: F401,F403
+from portainer_dashboard.services.portainer_utils import *  # noqa: F401,F403
+
+# Preserve existing __all__ for backward compatibility
+from portainer_dashboard.services.portainer_http import __all__ as _http_all
+from portainer_dashboard.services.portainer_normalizers import __all__ as _norm_all
+from portainer_dashboard.services.portainer_utils import __all__ as _util_all
+
+__all__ = [*_http_all, *_norm_all, *_util_all]
