+++ src/portainer_dashboard/websocket/remediation.py
@@ -10,59 +10,16 @@ from typing import Any
 
 from fastapi import APIRouter, WebSocket, WebSocketDisconnect
 
-from portainer_dashboard.auth.dependencies import SESSION_COOKIE_NAME
 from portainer_dashboard.config import get_settings
-from portainer_dashboard.core.session import SessionStorage
-from portainer_dashboard.dependencies import get_session_storage
-from portainer_dashboard.models.auth import SessionData
+from portainer_dashboard.core.audit import get_audit_logger
+from portainer_dashboard.core.ws_auth import authenticate_websocket
+from portainer_dashboard.core.ws_limiter import get_ws_tracker
 from portainer_dashboard.models.remediation import RemediationAction
 from portainer_dashboard.services.actions_store import get_actions_store
 
 LOGGER = logging.getLogger(__name__)
 
-
-async def _authenticate_websocket(websocket: WebSocket) -> SessionData | None:
-    """Authenticate a WebSocket connection using session cookie.
-
-    Args:
-        websocket: The WebSocket connection to authenticate.
-
-    Returns:
-        SessionData if authenticated, None otherwise.
-    """
-    token = websocket.cookies.get(SESSION_COOKIE_NAME)
-    if not token:
-        return None
-
-    storage: SessionStorage = get_session_storage()
-    record = storage.retrieve(token)
-    if record is None:
-        return None
-
-    now = datetime.now(timezone.utc)
-    settings = get_settings()
-
-    session_data = SessionData(
-        token=record.token,
-        username=record.username,
-        auth_method=record.auth_method,
-        authenticated_at=record.authenticated_at,
-        last_active=record.last_active,
-        session_timeout=record.session_timeout or settings.auth.session_timeout,
-    )
-
-    if session_data.is_expired(now):
-        storage.delete(token)
-        return None
-
-    # Update last active time
-    storage.touch(
-        token,
-        last_active=now,
-        session_timeout=session_data.session_timeout,
-    )
-
-    return session_data
+_audit = get_audit_logger()
 
 router = APIRouter()
 
@@ -201,13 +158,28 @@ async def remediation_websocket(websocket: WebSocket) -> None:
     - Request action summary statistics
     """
     # Authenticate before accepting connection
-    user = await _authenticate_websocket(websocket)
+    user = await authenticate_websocket(websocket)
     if user is None:
+        client_ip = websocket.client.host if websocket.client else "unknown"
+        _audit.log_websocket_rejected(client_ip, "/ws/remediation", "not_authenticated")
         await websocket.close(code=4001, reason="Not authenticated")
         LOGGER.warning("Remediation WebSocket connection rejected: not authenticated")
         return
 
+    # Check connection limit
+    tracker = get_ws_tracker()
+    if not tracker.can_connect(user.username):
+        _audit.log_websocket_rejected(
+            websocket.client.host if websocket.client else "unknown",
+            "/ws/remediation",
+            "connection_limit_exceeded",
+        )
+        await websocket.close(code=4029, reason="Too many connections")
+        return
+
     await websocket.accept()
+    tracker.on_connect(user.username, websocket)
+    _audit.log_websocket_connect(user.username, "/ws/remediation")
     LOGGER.info("Remediation WebSocket client connected for user: %s", user.username)
 
     async with _clients_lock:
@@ -252,6 +224,7 @@ async def remediation_websocket(websocket: WebSocket) -> None:
         except asyncio.CancelledError:
             pass
 
+        tracker.on_disconnect(user.username, websocket)
         async with _clients_lock:
             _connected_clients.discard(websocket)
             remaining = len(_connected_clients)
